control_ws/requirements.txt
agno>=2.0.0
redis>=4.5.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
PyYAML>=6.0
prometheus-client>=0.16.0
psutil>=5.9.0
